Add files via upload
Instructions
There are two parts to this assignment, complete both for full points.

30 Points
You are trying to go on a week long vacation. Sadly you are a student and money is tight. Here are the options you have seen which may work.

Going to Florence, Italy: plane ticket $1500
With a hotel stay at $150 a night
With a hostel stay at $20 a night, but with potentially noisy neighbors, and little privacy.
Staycation: $200 a night in a nearby resort
Camping in a park, $20 a night, but $100 of gas.
Visiting your parents. $0, and $100 of gas
Kayaking in a remote lake. $100 of gas, no parents, no drama, no staying costs, but mosquito galore
Design a program in pseudocode using several If-Then-Else which will:

ask how much money the user has, and then questions such as:
“Do you want to stay in a hotel?” or
“Do you want to go abroad?”
feel free to be as imaginative as you want in the questions you ask or even change the options. You can use a single character answer ‘y’ for yes or ‘n’ for no.
state how much the vacation will cost.
state whether the user will have enough money.
You don't need to stop the user midway through the questions. So, if they have big dreams but an empty wallet, you tell them that they don't have enough money.

If you do change the options, please summarize them at the top of the file. The number of options needs to be the same and one of them must have a choice inside, like the option to go to Florence.